- skip test in the _before method - remove CloneFieldTest (needs special treatment later) - remove get_expected_field_resolve_type from fields that haven't been worked on yet

+++ tests/wpunit/FieldTypes/NumberFieldTest.php
@@ -22,7 +22,7 @@ class NumberFieldTest extends \Tests\WPGraphQLAcf\WPUnit\AcfFieldTestCase {
 	}
 
 	public function get_expected_field_resolve_type(): ?string {
-		return 'Int';
+		return 'Float';
 	}
 
 }
